refactor(Lotto): `WinningNumber` 삭제, `Lotto`로 통합
