+++ src/main/java/lotto/controller/MainController.java
@@ -4,9 +4,9 @@ import java.util.List;
 import java.util.stream.Collectors;
 import lotto.controller.subcontroller.IssueLottoController;
 import lotto.domain.BonusNumber;
-import lotto.domain.WinningNumber;
+import lotto.domain.Lotto;
 import lotto.domain.repository.BonusNumberRepository;
-import lotto.domain.repository.WinningNumberRepository;
+import lotto.domain.repository.WinningLottoRepository;
 import lotto.view.InputView;
 import lotto.view.OutputView;
 
@@ -29,13 +29,13 @@ public class MainController {
         issueLottoController.process();
         {
             List<Integer> winnings = inputView.inputWinningNumber();
-            WinningNumberRepository.add(new WinningNumber(winnings));
+            WinningLottoRepository.add(new Lotto(winnings));
 
             int bonus = inputView.inputBonusNumber();
             BonusNumberRepository.add(new BonusNumber(bonus));
         }
         {
-            List<Integer> winnings = WinningNumberRepository.findWinningNumbers();
+            List<Integer> winnings = WinningLottoRepository.findWinningNumbers();
             System.out.println("\n당첨 번호: " + String.join(
                     ",", winnings.stream()
                             .map(String::valueOf)
