+++ src/main/java/lotto/domain/BonusNumber.java
@@ -1,9 +1,5 @@
 package lotto.domain;
 
-import java.util.List;
-import lotto.domain.repository.WinningNumberRepository;
-import lotto.util.ExceptionMessage;
-
 public class BonusNumber {
     private final int number;
