--- src/main/java/lotto/domain/WinningNumber.java
@@ -1,16 +1,0 @@
-package lotto.domain;
-
-import java.util.List;
-import lotto.util.ExceptionMessage;
-
-public class WinningNumber {
-    private final List<Integer> numbers;
-
-    public WinningNumber(List<Integer> numbers) {
-        this.numbers = numbers;
-    }
-
-    public List<Integer> getNumbers() {
-        return numbers;
-    }
-}
